+++ plugin/telegram/attachment.go
@@ -5,7 +5,7 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/usememos/memos/common/log"
+	"github.com/usememos/memos/internal/log"
 )
 
 type Attachment struct {
