+++ server/embed_frontend.go
@@ -8,7 +8,7 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
 
-	"github.com/usememos/memos/common/util"
+	"github.com/usememos/memos/internal/util"
 )
 
 //go:embed dist
